life slowly slipping away dealing with C++ types

+++ src/Mince.cpp
@@ -124,8 +124,8 @@ void bucketReads(sequence_parser* parser, std::atomic<uint64_t>* total,
 
             size_t readLength{s.length()};
             std::map<uint32_t, std::tuple<uint8_t, Direction>> merOffsetMap;
-            std::unordered_set<uint8_t> sforward = BucketModel::readHash(s, 8, false);
-            std::unordered_set<uint8_t> sreverse = BucketModel::readHash(s, 8, true);
+            std::unordered_set<uint16_t> sforward = BucketModel::readHash(s, 8, false);
+            std::unordered_set<uint16_t> sreverse = BucketModel::readHash(s, 8, true);
 
             size_t offset{0};
             size_t roffset{readLength};
@@ -260,8 +260,8 @@ void reassignOnsies(
         auto& bs = std::get<1>(kr);
         std::string s = bs.str;
         // pre-compute the hash of the minimers of s and rc(s)
-        std::unordered_set<uint8_t> sforward = BucketModel::readHash(s, 8, false);
-        std::unordered_set<uint8_t> sreverse = BucketModel::readHash(s, 8, true);
+        std::unordered_set<uint16_t> sforward = BucketModel::readHash(s, 8, false);
+        std::unordered_set<uint16_t> sreverse = BucketModel::readHash(s, 8, true);
 
         std::vector<uint8_t>& nlocs = bs.nlocs;
         size_t readLength{s.length()};
